Add Chapter 8 Example 3
This commit provides an example that performs direct convolutions using
VNNI instructions.  The Intel Optimization Manual already contains a
VNNI based direct convolution example (Chapter 8 Example 3) that is not in
this repository.  The existing assembly language example will be replaced
by the example in this commit in the next version of the Optimization
Manual.  This commit will not be merged until this new version of the
manual is released, so that the github repo and the manual remain in sync.
However, as the code is ready, it's being published in advance.

Co-authored-by: etay.meiri@intel.com

Signed-off-by: Mark Ryan <mark.d.ryan@intel.com>